% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endeavr-package.R
\docType{package}
\name{endeavr-package}
\alias{endeavr}
\alias{endeavr-package}
\title{endeavr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arthurhughes27/endeavr}
  \item \url{https://arthurhughes27.github.io/endeavr/}
  \item Report bugs at \url{https://github.com/arthurhughes27/endeavr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
