% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R, R/mvnpdfsmart.R
\name{mvnpdf}
\alias{mvnpdf}
\alias{mvnpdfsmart}
\title{Multivariate normal distribution probability density function
compute the value of the density of a multivariate normal distribution on at n points.}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix, with n columns (the observations) and p rows}

\item{mean}{a vector of means}

\item{varcovM}{a variance-covariance matrix}

\item{Log}{a logical parameter, zith default TRUE, which takes the log of the pdf}
}
\value{
Returns a list containing the matrix x, and a vector of length n of
the multivariate normal distribution density values at those points
}
\description{
Multivariate normal distribution probability density function
compute the value of the density of a multivariate normal distribution on at n points.
}
